+++ lib/fundamental/icons.dart
@@ -6,19 +6,21 @@ class ThemedIcon extends StatelessWidget {
 
   final IconData icon;
   final Key key;
-  final IconSize size;
+  final IconSize iconSize;
   final ThemeGroupType type;
   final String semanticLabel;
   final TextDirection textDirection;
+  final Emphasis emphasis;
 
   ThemedIcon(
       this.icon,
       {
         this.type  = ThemeGroupType.MOM,
         this.key,
-        this.size = IconSize.SMALL,
+        this.iconSize = IconSize.SMALL,
         this.semanticLabel,
-        this.textDirection
+        this.textDirection,
+        this.emphasis = Emphasis.NONE
       }
   );
 
@@ -28,17 +30,60 @@ class ThemedIcon extends StatelessWidget {
     //construct a default widget in order to fall back on default values when the optional params aren't passed
     var defaultIcon = Icon(icon);
 
-    return Consumer<MaterialThemesManager> (
-      builder: (context, themeManager, child) {
-        return Icon(
-          icon,
-          key: key != null ? key : defaultIcon.key,
-          size: themeManager.getTheme(type, iconSize: size).iconTheme.size,
-          color: themeManager.getTheme(type).iconTheme.color,
-          semanticLabel: semanticLabel != null ? semanticLabel : defaultIcon.semanticLabel,
-          textDirection: textDirection != null ? textDirection : defaultIcon.textDirection,
-        );
-      },
+    return Icon(
+      icon,
+      key: key != null ? key : defaultIcon.key,
+      size: context.watch<MaterialThemesManager>().getTheme(type, iconSize: iconSize).iconTheme.size,
+      color: context.watch<MaterialThemesManager>().getTheme(type, emphasis: emphasis).iconTheme.color,
+      semanticLabel: semanticLabel != null ? semanticLabel : defaultIcon.semanticLabel,
+      textDirection: textDirection != null ? textDirection : defaultIcon.textDirection,
+    );
+  }
+}
+
+class ThemedIconButton extends StatelessWidget {
+
+  final IconData icon;
+  final Key key;
+  final IconSize iconSize;
+  final ThemeGroupType type;
+  final String semanticLabel;
+  final TextDirection textDirection;
+  final Function onPressedCallback;
+  final Emphasis emphasis;
+
+  ThemedIconButton(
+    this.icon,
+      {
+        this.type  = ThemeGroupType.MOM,
+        this.key,
+        this.iconSize = IconSize.SMALL,
+        this.semanticLabel,
+        this.textDirection,
+        this.onPressedCallback,
+        this.emphasis = Emphasis.NONE
+      }
+  );
+
+  @override
+  Widget build(BuildContext context) {
+
+    var buttonIcon = ThemedIcon(
+      icon,
+      key: key,
+      iconSize: iconSize,
+      type: type,
+      semanticLabel: semanticLabel,
+      textDirection: textDirection,
+      emphasis: emphasis
+    );
+
+    return IconButton(
+      onPressed: onPressedCallback != null ? onPressedCallback : () => print("Tapped ${icon.toString()}"),
+      icon: buttonIcon,
+      // TODO iconSize: appbarIconSize,
+      // TODO color: iconDarkBgColor,
+      // TODO padding: EdgeInsets.only(right: appbarIconMargin),
     );
   }
 }