+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## [0.0.1] - TODO: Add release date.
-
-* TODO: Describe initial release.
+## [0.0.3] - Added ThemedIconButton. Added Emphasis to ThemedIcon
+## [0.0.2] - Added fundamentals, example, login screen
+## [0.0.1] - Initial release.